dockerfiles/pythonscripts/requirements_consumer.txt
kafka-python
mysql-connector-python

